refactor: code style and improvements
