+++ src/components/Select/__stories__/SelectShowcase.tsx
@@ -3,18 +3,16 @@ import React from 'react';
 import {Plus, TrashBin} from '@gravity-ui/icons';
 import range from 'lodash/range';
 
-import {Select} from '..';
 import type {SelectOption, SelectProps} from '..';
+import {Select} from '..';
 import {Button} from '../../Button';
 import {ClipboardButton} from '../../ClipboardButton';
 import {Icon} from '../../Icon';
-import {RadioButton} from '../../RadioButton';
 import type {RadioButtonOption} from '../../RadioButton';
+import {RadioButton} from '../../RadioButton';
 import {Tooltip} from '../../Tooltip';
 import {TextInput} from '../../controls';
 import {block} from '../../utils/cn';
-import {generateOptions} from '../__tests__/utils';
-import {useSelectAllFilter} from '../hooks';
 
 import {
     EXAMPLE_CHILDREN_OPTIONS,
@@ -32,7 +30,7 @@ import {
 import './SelectShowcase.scss';
 
 const b = block('select-showcase');
-const Mode = {
+export const Mode = {
     CODE: 'code',
     VIEW: 'view',
 };
@@ -48,7 +46,7 @@ const getEscapedString = (str: string) => {
     return str.replace(/[-/\\^$*+?.()|[\]{}]/g, '\\$&');
 };
 
-const radioButtonOptions: RadioButtonOption[] = [
+export const radioButtonOptions: RadioButtonOption[] = [
     {value: Mode.VIEW, content: 'View'},
     {value: Mode.CODE, content: 'Code'},
 ];
@@ -91,80 +89,6 @@ const ExampleItem = (props: {
                     {...selectProps}
                     value={value}
                     onUpdate={(nextValue) => setValue(nextValue)}
-                >
-                    {children}
-                </Select>
-            ) : (
-                code.map((codeItem, i) => {
-                    return (
-                        <React.Fragment key={`${title}-${i}`}>
-                            <pre>
-                                {codeItem}
-                                <ClipboardButton
-                                    className={b('copy-button')}
-                                    size={16}
-                                    text={codeItem}
-                                />
-                            </pre>
-                        </React.Fragment>
-                    );
-                })
-            )}
-        </div>
-    );
-};
-
-const ExampleSelectAllItem = (props: {
-    title: string;
-    selectProps: SelectProps;
-    code?: string[];
-    children?: SelectProps['children'];
-}) => {
-    const {title, selectProps, children, code = []} = props;
-    const multiple = props.selectProps.multiple;
-    const [mode, setMode] = React.useState(Mode.VIEW);
-    const [value, setValue] = React.useState<string[]>([]);
-
-    React.useEffect(() => {
-        if (!multiple) {
-            setValue([]);
-        }
-    }, [multiple]);
-
-    const onUpdate = (nextValue: string[]) => setValue(nextValue);
-
-    const renderFilter = useSelectAllFilter({
-        value: value,
-        options: selectProps.options,
-        onUpdate: onUpdate,
-        hasClear: true,
-        autoFocus: true,
-    });
-
-    return (
-        <div className={b('example-item')}>
-            <h3>
-                {title}
-                {Boolean(code.length) && (
-                    <RadioButton
-                        className={b('example-item-radio')}
-                        size="s"
-                        value={mode}
-                        onUpdate={(nextMode) => setMode(nextMode)}
-                    >
-                        <RadioButton.Option {...radioButtonOptions[0]} />
-                        <RadioButton.Option {...radioButtonOptions[1]} />
-                    </RadioButton>
-                )}
-            </h3>
-            {mode === Mode.VIEW ? (
-                <Select
-                    {...selectProps}
-                    value={value}
-                    onUpdate={onUpdate}
-                    renderFilter={renderFilter}
-                    filterable
-                    multiple
                 >
                     {children}
                 </Select>
@@ -456,19 +380,6 @@ export const SelectShowcase = (props: SelectProps) => {
                 <Select.Option value="val3" content="Value3" />
                 <Select.Option value="val4" content="Value4" />
             </ExampleItem>
-            <ExampleSelectAllItem
-                title="Select with select all hook"
-                code={[EXAMPLE_JSON_OPTIONS]}
-                selectProps={{
-                    ...props,
-                    options: generateOptions([
-                        ['val1', 'val1'],
-                        ['val11', 'val11'],
-                        ['val2', 'val2'],
-                        ['val3', 'val3'],
-                    ]),
-                }}
-            />
         </div>
     );
 };
