+++ src/components/Select/hooks/index.ts
@@ -1,2 +1,1 @@
 export {useQuickSearch} from './useQuickSearch';
-export {useSelectAllFilter} from './useSelectAllFilter/useSelectAllFilter';
