--- src/components/Select/hooks/useSelectAllFilter/useSelectAllFilter.scss
@@ -1,10 +1,0 @@
-.select-filter-with-all {
-    display: flex;
-    flex-direction: row;
-    gap: 0.5rem;
-    align-items: center;
-
-    &__no-options {
-        width: calc(100% - 8px);
-    }
-}
