+++ src/hooks/index.ts
@@ -9,6 +9,7 @@ export * from './useListNavigation';
 export * from './useOutsideClick';
 export * from './usePortalContainer';
 export * from './useSelect';
+export * from './useSelectAllFilter';
 export * from './useTimeout';
 export * from './useViewportSize';
 export * from './useVirtualElementRef';
