+++ src/hooks/useSelectAllFilter/i18n/en.json
@@ -1,0 +1,4 @@
+{
+  "label_select_all": "Select all",
+  "label_remove_selection": "Remove selection"
+}
