+++ src/hooks/useSelectAllFilter/i18n/index.ts
@@ -1,0 +1,8 @@
+import {addComponentKeysets} from '../../../components/utils/addComponentKeysets';
+
+import en from './en.json';
+import ru from './ru.json';
+
+const COMPONENT = 'Select all hook';
+
+export default addComponentKeysets({en, ru}, COMPONENT);
