+++ src/hooks/useSelectAllFilter/i18n/ru.json
@@ -1,0 +1,4 @@
+{
+  "label_select_all": "Выбрать все",
+  "label_remove_selection": "Снять выделение"
+}
